+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@ This lets us remove a lot of code that is just there to cater to the needs of ol
 
 * `ActionToggle` is removed - just use `onChange(of:)` to observe the toggle state instead.
 * `FilePicker` now requires `UTType` instead of `String`.
+* `ScanCodeGenerator` is replaced by `Image` and `ImageRepresentable` init extensions.